refactor: SeismicityDepthPoint モデルを専用ファイルへ分離
