+++ app/lib/feature/seismicity/data/model/seismicity_depth_point.dart
@@ -1,0 +1,15 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'seismicity_depth_point.freezed.dart';
+
+/// 深さ断面図の1点(投影後の軸値・深さ・マグニチュード)。
+@freezed
+abstract class SeismicityDepthPoint with _$SeismicityDepthPoint {
+  const factory SeismicityDepthPoint({
+    /// 投影軸の値(緯度 or 経度)
+    required double axisValue,
+    required double depth,
+    required double? magnitude,
+    required String eventId,
+  }) = _SeismicityDepthPoint;
+}
